agent/requirements.txt
# UniGPU GPU Agent — Python Dependencies

websockets>=12.0
docker>=7.0
httpx>=0.27
python-dotenv>=1.0
pynvml>=12.0
